Igor

+++ PHP/actions/lists/ajax_list_edit.php
@@ -11,10 +11,12 @@
 if (isset($_GET['function']) && $_GET['function'] == 'validUser') {
     if ($_GET['username'] == $_SESSION['username']) {
       echo -1;
-    } else if ($listsDB->isUserOnList($_GET['id'], $_GET['username'])){
-      echo -3;
-    } elseif ($userDB->checkUsernameExists($_GET['username'])) {
-      $listsDB->addUser($_GET['id'], $_GET['username']);
+    } if ($listsDB->isUserOnList($_GET['id'], $_GET['username'])){
+          echo -3;
+    } else if ($userDB->checkUsernameExists($_GET['username'])) {
+      if (isset($_GET['id']) && ($_GET['id'] != -1)) {
+        $listsDB->addUser($_GET['id'], $_GET['username']);
+      }
       echo 0;
     }  else {
       echo -2;
@@ -22,26 +24,36 @@ if (isset($_GET['function']) && $_GET['function'] == 'validUser') {
 }
 
 if (isset($_GET['function']) && $_GET['function'] == 'distinctItem') {
+  if (isset($_GET['id']) && ($_GET['id']) != -1) {
     if ($listsDB->itemExists($_GET['id'], $_GET['description'])) {
       echo -1;
     } else {
         $listsDB->addItem($_GET['id'], $_GET['description']);
       echo 0;
     }
+  } else {
+    echo 0;
+  }
 }
 
 if (isset($_GET['function']) && $_GET['function'] == 'removeItem') {
-    $listsDB->removeItem($_GET['id'], $_GET['description']);
+    if (isset($_GET['id']) && ($_GET['id']) != -1) {
+      $listsDB->removeItem($_GET['id'], $_GET['description']);
+    }
     echo 0;
 }
 
 if (isset($_GET['function']) && $_GET['function'] == 'removeUser') {
-    $listsDB->removeUser($_GET['id'], $_GET['username']);
+    if (isset($_GET['id']) && ($_GET['id']) != -1) {
+      $listsDB->removeUser($_GET['id'], $_GET['username']);
+    }
     echo 0;
 }
 
 if (isset($_GET['function']) && $_GET['function'] == 'updateDone') {
-    $listsDB->updateDone($_GET['id'], $_GET['description'], $_GET['value']);
+    if (isset($_GET['id']) && ($_GET['id']) != -1) {
+      $listsDB->updateDone($_GET['id'], $_GET['description'], $_GET['value']);
+    }
     echo 0;
 }
 ?>
